requirements.txt
viam-sdk
opencv-python-headless
Pillow
numpy
